+++ backend/src/main/java/com/synergyhub/dto/request/UpdateRoleRequest.java
@@ -23,5 +23,5 @@ public class UpdateRoleRequest {
     private String description;
     
     @Builder.Default
-    private Set<Integer> permissionIds = Set.of();
+    private Set<Long> permissionIds = Set.of();
 }
