+++ backend/src/main/java/com/synergyhub/dto/response/ProjectDetailResponse.java
@@ -14,10 +14,10 @@ import java.util.List;
 @AllArgsConstructor
 public class ProjectDetailResponse {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer organizationId;
+    private Long organizationId;
     private String organizationName;
     private ProjectLeadResponse projectLead;
     private LocalDate startDate;
