+++ backend/src/main/java/com/synergyhub/dto/response/ProjectResponse.java
@@ -13,10 +13,10 @@ import java.time.LocalDate;
 @AllArgsConstructor
 public class ProjectResponse {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer organizationId;
+    private Long organizationId;
     private String organizationName;
     private ProjectLeadResponse projectLead;
     private LocalDate startDate;
