+++ backend/src/main/java/com/synergyhub/events/task/TaskCreatedEvent.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 public class TaskCreatedEvent extends TaskEvent {
     @Getter private final String projectName; // Extra context often needed for emails
 
-    public TaskCreatedEvent(User actor, String ipAddress, Integer projectId, String projectName, String taskTitle) {
+    public TaskCreatedEvent(User actor, String ipAddress, Long projectId, String projectName, String taskTitle) {
         super(actor, ipAddress, projectId, projectName, taskTitle);
         this.projectName = projectName;
     }
