+++ backend/src/main/java/com/synergyhub/events/task/TaskEvent.java
@@ -7,14 +7,14 @@ import lombok.Getter;
 
 @Getter
 public abstract class TaskEvent extends BaseEvent {
-    private final Integer projectId;  // Loose reference
+    private final Long projectId;  // Loose reference
     private final String taskTitle;
     
     // Optional: You can include projectName if you have it cheap, 
     // useful for email subjects without a DB lookup.
     private final String projectName; 
 
-    protected TaskEvent(User actor, String ipAddress, Integer projectId, String projectName, String taskTitle) {
+    protected TaskEvent(User actor, String ipAddress, Long projectId, String projectName, String taskTitle) {
         super(actor, ipAddress);
         this.projectId = projectId;
         this.projectName = projectName;
