+++ backend/src/main/java/com/synergyhub/events/task/TaskUpdatedEvent.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class TaskUpdatedEvent extends TaskEvent {
     private final String changes;
 
-    public TaskUpdatedEvent(User actor, String ipAddress, Integer projectId, String projectName, String taskTitle, String changes) {
+    public TaskUpdatedEvent(User actor, String ipAddress, Long projectId, String projectName, String taskTitle, String changes) {
         super(actor, ipAddress, projectId, projectName, taskTitle);
         this.changes = changes;
     }
