+++ backend/src/main/java/com/synergyhub/exception/ProjectNameAlreadyExistsException.java
@@ -2,7 +2,7 @@ package com.synergyhub.exception;
 
 public class ProjectNameAlreadyExistsException extends BadRequestException {
 
-    public ProjectNameAlreadyExistsException(String projectName, Integer organizationId) {
+    public ProjectNameAlreadyExistsException(String projectName, Long organizationId) {
         super("Project with name '" + projectName + "' already exists in organization " + organizationId);
     }
 }
