+++ backend/src/main/java/com/synergyhub/exception/ProjectNotFoundException.java
@@ -2,11 +2,11 @@ package com.synergyhub.exception;
 
 public class ProjectNotFoundException extends ResourceNotFoundException {
 
-    public ProjectNotFoundException(Integer projectId) {
+    public ProjectNotFoundException(Long projectId) {
         super("Project", "id", projectId);
     }
 
-    public ProjectNotFoundException(String projectName, Integer organizationId) {
+    public ProjectNotFoundException(String projectName, Long organizationId) {
         super("Project not found with name '" + projectName + "' in organization " + organizationId);
     }
 }
