+++ backend/src/main/java/com/synergyhub/exception/UnauthorizedRoleManagementException.java
@@ -2,7 +2,7 @@ package com.synergyhub.exception;
 
 public class UnauthorizedRoleManagementException extends RuntimeException {
 
-    public UnauthorizedRoleManagementException(Integer organizationId) {
+    public UnauthorizedRoleManagementException(Long organizationId) {
         super("User is not an admin of organization " + organizationId + " and cannot manage roles");
     }
