+++ backend/src/main/java/com/synergyhub/repository/PermissionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
     
     Optional<Permission> findByName(String name);
