+++ backend/src/main/java/com/synergyhub/repository/ProjectMemberRepository.java
@@ -14,36 +14,36 @@ import java.util.Optional;
 public interface ProjectMemberRepository extends JpaRepository<ProjectMember, ProjectMemberId> {
 
     // Find all members of a project
-    List<ProjectMember> findByProjectId(Integer projectId);
+    List<ProjectMember> findByProjectId(Long projectId);
 
-    List<ProjectMember> findByUserId(Integer userId);
+    List<ProjectMember> findByUserId(Long userId);
 
-    boolean existsByProjectIdAndUserId(Integer projectId, Integer userId);
+    boolean existsByProjectIdAndUserId(Long projectId, Long userId);
 
     @Query("SELECT pm FROM ProjectMember pm WHERE pm.project.id = :projectId AND pm.user.id = :userId")
-    Optional<ProjectMember> findByProjectIdAndUserId(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    Optional<ProjectMember> findByProjectIdAndUserId(@Param("projectId") Long projectId, @Param("userId") Long userId);
 
-    List<ProjectMember> findByProjectIdAndRole(Integer projectId, String role);
+    List<ProjectMember> findByProjectIdAndRole(Long projectId, String role);
 
-    void deleteByProjectIdAndUserId(Integer projectId, Integer userId);
+    void deleteByProjectIdAndUserId(Long projectId, Long userId);
 
-    long countByProjectId(Integer projectId);
+    long countByProjectId(Long projectId);
 
     @Query("SELECT CASE WHEN COUNT(pm) > 0 THEN true ELSE false END " +
             "FROM ProjectMember pm WHERE pm.project.id = :projectId AND pm.user.id = :userId")
-    boolean hasAccessToTaskProject(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    boolean hasAccessToTaskProject(@Param("projectId") Long projectId, @Param("userId") Long userId);
 
     @Query("SELECT CASE WHEN COUNT(p) > 0 THEN true ELSE false END " +
             "FROM Project p WHERE p.id = :projectId AND p.projectLead.id = :userId")
-    boolean isProjectLead(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    boolean isProjectLead(@Param("projectId") Long projectId, @Param("userId") Long userId);
 
     @Query("SELECT pm.role FROM ProjectMember pm WHERE pm.project.id = :projectId AND pm.user.id = :userId")
-    Optional<String> getUserRoleInProject(@Param("projectId") Integer projectId, @Param("userId") Integer userId);
+    Optional<String> getUserRoleInProject(@Param("projectId") Long projectId, @Param("userId") Long userId);
 
     @Query("SELECT CASE WHEN COUNT(pm) > 0 THEN true ELSE false END " +
             "FROM ProjectMember pm WHERE pm.project.id = :projectId AND pm.user.id = :userId " +
             "AND pm.role IN :allowedRoles")
-    boolean hasRoleInProject(@Param("projectId") Integer projectId,
-                             @Param("userId") Integer userId,
+    boolean hasRoleInProject(@Param("projectId") Long projectId,
+                             @Param("userId") Long userId,
                              @Param("allowedRoles") List<String> allowedRoles);
 }
