+++ backend/src/main/java/com/synergyhub/security/JwtTokenProvider.java
@@ -62,7 +62,7 @@ public class JwtTokenProvider {
         return generateTokenFromUserId(userPrincipal.getId(), userPrincipal.getEmail());
     }
 
-    public String generateTokenFromUserId(Integer userId, String email) {
+    public String generateTokenFromUserId(Long userId, String email) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + jwtExpirationMs);
 
@@ -76,7 +76,7 @@ public class JwtTokenProvider {
                 .compact();
     }
 
-    public String generateTemporaryToken(Integer userId, String email) {
+    public String generateTemporaryToken(Long userId, String email) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + 300000); // 5 minutes
 
@@ -91,7 +91,7 @@ public class JwtTokenProvider {
                 .compact();
     }
 
-    public String generateRefreshToken(Integer userId, String email) {
+    public String generateRefreshToken(Long userId, String email) {
         Date now = new Date();
         Date expiryDate = new Date(now.getTime() + jwtExpirationMs * 7); // 7 days (or longer as needed)
 
@@ -106,14 +106,14 @@ public class JwtTokenProvider {
                 .compact();
     }
 
-    public Integer getUserIdFromToken(String token) {
+    public Long getUserIdFromToken(String token) {
         Claims claims = Jwts.parser()
                 .verifyWith(getSigningKey())
                 .build()
                 .parseSignedClaims(token)
                 .getPayload();
 
-        return Integer.parseInt(claims.getSubject());
+        return Long.parseLong(claims.getSubject());
     }
 
     public String getEmailFromToken(String token) {
